# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/create-user": {
            "post": {
                "description": "Criar um usuario a partir das novas infos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Criar um novo usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email do usuario",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nome do usuario",
                        "name": "Name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Endereço do usuario",
                        "name": "Adress",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/edit-adress": {
            "put": {
                "description": "Editar um usuario a partir das novas infos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Editar usuario (Password)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Password para editar as novas info",
                        "name": "Password",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email para editar as novas info",
                        "name": "Email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/edit-email": {
            "put": {
                "description": "Editar um usuario a partir das novas infos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Editar usuario (Email)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email para editar as novas info",
                        "name": "Email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/edit-name": {
            "put": {
                "description": "Editar um usuario a partir das novas infos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Editar usuario (Name)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Name para editar as novas info",
                        "name": "Name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Email para editar as novas info",
                        "name": "Email",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/get-user": {
            "get": {
                "description": "Buscar um usuario dentro da db",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Pegar um usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email para pegar um usuario especifico",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/loggout": {
            "put": {
                "description": "Altera o modo de login para false",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Desconectar o usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email para pegar o atributo do login",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/login": {
            "put": {
                "description": "Altera o modo de login para true",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Login"
                ],
                "summary": "Conectar o usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email para pegar o atributo do login",
                        "name": "Email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/all-order": {
            "get": {
                "description": "Obtém todas as ordens com base nos parâmetros de pesquisa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Obter todas as ordens",
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/all-restaurant": {
            "get": {
                "description": "Encontra todos os restaurantes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Buscar todos restaurante",
                "responses": {
                    "200": {
                        "description": "{\"message\": \"result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/create-order": {
            "post": {
                "description": "Cria uma nova ordem com base nos parâmetros fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Criar uma nova ordem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id do restaurante que criou a ordem de pedido pro usuario",
                        "name": "RestaurantID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "array",
                        "description": "Comidas/Pedido",
                        "name": "Items",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/create-restaurant": {
            "post": {
                "description": "Criação de novos restaurantes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Criar novo restaurante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do Restaurante",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Restaurante criado\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/find-order": {
            "get": {
                "description": "Obtém uma ordem com base nos parâmetros de pesquisa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Obter uma ordem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Numero do pedido",
                        "name": "orderNumber",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/find-restaurant": {
            "get": {
                "description": "Encontra um restaurante",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Restaurant"
                ],
                "summary": "Buscar um restaurante",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do Restaurante",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/food": {
            "get": {
                "description": "Encontra alimentos com base nos parâmetros de pesquisa",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Food"
                ],
                "summary": "Encontrar alimentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do alimento",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Descrição do alimento",
                        "name": "description",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Alimentos encontrados\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Criar novos alimentos :)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Food"
                ],
                "summary": "Criar alimentos",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do alimento",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Descrição do alimento",
                        "name": "description",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Valor do Produto",
                        "name": "price",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Registro adicionado com sucesso\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/status-order": {
            "put": {
                "description": "Altera o status de uma ordem com base nos parâmetros fornecidos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Order"
                ],
                "summary": "Alterar o status de uma ordem",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nome do alimento",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Descrição do alimento",
                        "name": "description",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"message\": \"Result\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}